+++ e2e/instance/instance.go
@@ -10,7 +10,9 @@
 package instance
 
 import (
+	"bufio"
 	"bytes"
+	"fmt"
 	"os"
 	"os/user"
 	"path/filepath"
@@ -18,6 +20,7 @@ import (
 	"strings"
 	"syscall"
 	"testing"
+	"time"
 
 	"github.com/apptainer/apptainer/e2e/internal/e2e"
 	"github.com/apptainer/apptainer/e2e/internal/testhelper"
@@ -134,7 +137,7 @@ func (c *ctx) testCreateManyInstances(t *testing.T) {
 			e2e.ExpectExit(0),
 		)
 
-		c.expectInstance(t, instanceName, 1)
+		c.expectInstance(t, instanceName, 1, false)
 	}
 }
 
@@ -383,7 +386,7 @@ func (c *ctx) testInstanceWithConfigDir(t *testing.T) {
 		e2e.ExpectExit(0),
 	)
 
-	c.expectInstance(t, name, 1)
+	c.expectInstance(t, name, 1, false)
 	c.execInstance(t, name, "id")
 	c.stopInstance(t, name)
 
@@ -392,6 +395,58 @@ func (c *ctx) testInstanceWithConfigDir(t *testing.T) {
 			t.Fatalf("failed %v", err)
 		}
 	})(t)
+}
+
+// testShareNSMopde will test --sharens flag
+func (c *ctx) testShareNSMode(t *testing.T) {
+	dir, err := os.MkdirTemp(c.env.TestDir, "InstanceShareNS")
+	if err != nil {
+		t.Fatalf("Failed to create temporary directory: %v", err)
+	}
+	defer os.RemoveAll(dir)
+
+	file := fmt.Sprintf("%s/file", dir)
+	f, err := os.Create(file)
+	if err != nil {
+		t.Fatalf("failed to create file, this is unexpected, err: %v", err)
+	}
+	f.Close()
+
+	insNumber := 2
+	for i := 0; i < insNumber; i++ {
+		go c.env.RunApptainer(
+			t,
+			e2e.WithProfile(c.profile),
+			e2e.WithCommand("exec"),
+			e2e.WithArgs("--bind", fmt.Sprintf("%s:/canary/file", file), "--sharens", c.env.ImagePath, "sh", "-c", "echo 0 >> /canary/file; sleep 1"),
+			e2e.ExpectExit(0),
+		)
+	}
+
+	// waiting enough time for the file written
+	time.Sleep(2 * time.Second)
+
+	f, err = os.Open(file)
+	if err != nil {
+		t.Fatalf("failed to open file: %v, this is unexpected", err)
+	}
+	defer f.Close()
+
+	scanner := bufio.NewScanner(f)
+	scanner.Split(bufio.ScanLines)
+
+	var count int
+	for scanner.Scan() {
+		count++
+	}
+
+	if err := scanner.Err(); err != nil {
+		t.Fatalf("having issue while scanning file: %s, err: %v", file, err)
+	}
+
+	if count != insNumber {
+		t.Fatalf("should have %d lines, but actually is %d", insNumber, count)
+	}
 }
 
 // E2ETests is the main func to trigger the test suite
@@ -424,6 +479,7 @@ func E2ETests(env e2e.TestEnv) testhelper.Tests {
 				{"GhostInstance", c.testGhostInstance},
 				{"CheckpointInstance", c.testCheckpointInstance},
 				{"InstanceWithConfigDir", c.testInstanceWithConfigDir},
+				{"ShareNSMode", c.testShareNSMode},
 			}
 
 			profiles := []e2e.Profile{